requirements.txt
google-adk==1.0.0
openai
anthropic
faster-whisper
sentence-transformers
python-dotenv
ffmpeg-python
google-cloud-speech
google-cloud-storage
google-cloud-texttospeech
requests
beautifulsoup4
tweepy
wordpress-api
pydub
torch
transformers
PyYAML
structlog
pytest
mypy
black
ruff 